cleanup(headers): eliminate allreport.h
