+++ src/general_report_manager.cpp
@@ -25,6 +25,8 @@
 #include "paths.h"
 #include "platfdep.h"
 #include "util.h"
+#include "option.h"
+#include "reports/reportbase.h"
 
 #include "model/Model_Infotable.h"
 #include "model/Model_Report.h"
