+++ src/import_export/qif_import_gui.cpp
@@ -24,6 +24,7 @@ Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 #include "paths.h"
 #include "util.h"
 #include "webapp.h"
+#include "option.h"
 
 #include "model/Model_Category.h"
 #include "model/Model_Subcategory.h"
