+++ src/mmreportspanel.h
@@ -21,10 +21,14 @@
 #define MM_EX_REPORTSPANEL_H_
 
 #include "mmpanelbase.h"
-#include "reports/reportbase.h"
 class wxWebView;
 class mmGUIFrame;
 class mmDateRange;
+class wxChoice;
+class wxDatePickerCtrl;
+class wxDateEvent;
+class wxChoice;
+class mmPrintableBase;
 
 class mmReportsPanel : public mmPanelBase
 {
@@ -98,4 +102,3 @@ public:
 };
 
 #endif
-
