+++ src/option.cpp
@@ -23,13 +23,23 @@
 #include "constants.h"
 #include "images_list.h"
 #include "singleton.h"
+#include "maincurrencydialog.h"
 #include "model/Model_Infotable.h"
 #include "model/Model_Setting.h"
 #include "model/Model_Account.h"
-#include "maincurrencydialog.h"
 #include "model/Model_Currency.h"
 #include "model/Model_CurrencyHistory.h"
-#include "reports/allreport.h"
+#include "reports/myusage.h"
+#include "reports/summary.h"
+#include "reports/categexp.h"
+#include "reports/categovertimeperf.h"
+#include "reports/payee.h"
+#include "reports/incexpenses.h"
+#include "reports/budgetingperf.h"
+#include "reports/cashflow.h"
+#include "reports/budgetcategorysummary.h"
+#include "reports/summarystocks.h"
+#include "reports/forecast.h"
 
 struct ReportInfo
 {
