--- src/reports/allreport.h
@@ -1,36 +1,0 @@
-/*******************************************************
- Copyright (C) 2016 Guan Lisheng (guanlisheng@gmail.com)
-
- This program is free software; you can redistribute it and/or modify
- it under the terms of the GNU General Public License as published by
- the Free Software Foundation; either version 2 of the License, or
- (at your option) any later version.
-
- This program is distributed in the hope that it will be useful,
- but WITHOUT ANY WARRANTY; without even the implied warranty of
- MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- GNU General Public License for more details.
-
- You should have received a copy of the GNU General Public License
- along with this program; if not, write to the Free Software
- Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
- ********************************************************/
-
-#pragma once
-
-#include "budget.h"
-#include "budgetcategorysummary.h"
-#include "budgetingperf.h"
-#include "cashflow.h"
-#include "categexp.h"
-#include "categovertimeperf.h"
-#include "htmlbuilder.h"
-#include "incexpenses.h"
-#include "mmDateRange.h"
-#include "myusage.h"
-#include "payee.h"
-#include "reportbase.h"
-#include "summary.h"
-#include "summarystocks.h"
-#include "transactions.h"
-#include "forecast.h"
