+++ src/reports/categexp.cpp
@@ -19,9 +19,8 @@
 
 #include "categexp.h"
 #include "budget.h"
-#include "htmlbuilder.h"
-
-#include "htmlbuilder.h"
+#include "reports/htmlbuilder.h"
+#include "reports/mmDateRange.h"
 #include "option.h"
 #include <algorithm>
 #include "model/Model_Category.h"
