+++ src/reports/categovertimeperf.cpp
@@ -18,10 +18,12 @@ Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 
 //----------------------------------------------------------------------------
 #include "categovertimeperf.h"
-#include "htmlbuilder.h"
+#include "reports/htmlbuilder.h"
 #include "util.h"
-#include <algorithm>
+#include "option.h"
+#include "reports/mmDateRange.h"
 #include "model/Model_Category.h"
+#include <algorithm>
 #include <map>
 
 #define CATEGORY_SORT_BY_NAME       1
