+++ src/reports/forecast.cpp
@@ -17,6 +17,8 @@
  ********************************************************/
 
 #include "forecast.h"
+#include "option.h"
+#include "reports/mmDateRange.h"
 #include "model/Model_Checking.h"
 
 class mm_html_template;
