+++ src/reports/myusage.cpp
@@ -17,8 +17,10 @@ Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
 ********************************************************/
 
 #include "myusage.h"
-#include "model/Model_Usage.h"
+#include "option.h"
+#include "reports/mmDateRange.h"
 #include "reports/htmlbuilder.h"
+#include "model/Model_Usage.h"
 #include "model/Model_Report.h"
 
 const char *usage_template = R"(
