+++ src/reports/summarystocks.cpp
@@ -17,16 +17,17 @@
  ********************************************************/
 
 #include "summarystocks.h"
-#include "htmlbuilder.h"
+#include "reports/htmlbuilder.h"
 
 #include "constants.h"
 #include "stockspanel.h"
+#include "budget.h"
 #include "util.h"
+#include "reports/mmDateRange.h"
 #include "model/Model_Account.h"
 #include "model/Model_Currency.h"
 #include "model/Model_CurrencyHistory.h"
 #include "model/Model_StockHistory.h"
-#include "budget.h"
 
 #include <algorithm>
